petsync_backend/requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
pandas
reportlab
httpx
pytest
pytest-asyncio
requests
bcrypt
python-jose[cryptography]
